refactor: update kind 411 to kind 10411 throughout codebase
Change all references from kind 411 to kind 10411 in documentation,
code, imports, and function calls. Move kind411 handler to kind10411
directory to maintain consistency with the updated NIP specification.
